Change longs back to doubles and fix test

+++ server/src/main/java/com/tco/requests/DistancesRequest.java
@@ -29,28 +29,28 @@ public class DistancesRequest extends Request {
 
   public ArrayList<Integer> getDistances(Places places) {
     ArrayList<Integer> distances = new ArrayList<Integer>();
-    long latitude1 = 0;
-    long longitude1 = 0;
-    long latitude2 = 0;
-    long longitude2 = 0;
+    double latitude1 = 0;
+    double longitude1 = 0;
+    double latitude2 = 0;
+    double longitude2 = 0;
     for (int i = 1; i < places.size(); i++) {
-      latitude1 = Long.parseLong(places.get(i - 1).get("latitude"));
-      longitude1 = Long.parseLong(places.get(i - 1).get("longitude"));
+      latitude1 = Double.parseDouble(places.get(i - 1).get("latitude"));
+      longitude1 = Double.parseDouble(places.get(i - 1).get("longitude"));
 
-      latitude2 = Long.parseLong(places.get(i).get("latitude"));
-      longitude2 = Long.parseLong(places.get(i).get("longitude"));
+      latitude2 = Double.parseDouble(places.get(i).get("latitude"));
+      longitude2 = Double.parseDouble(places.get(i).get("longitude"));
 
       int distance =
           computeDistance(latitude1, latitude2, longitude1, longitude2);
       distances.add(distance);
     }
     latitude1 =
-        Long.parseLong(places.get(places.size() - 1).get("latitude"));
+        Double.parseDouble(places.get(places.size() - 1).get("latitude"));
     longitude1 =
-        Long.parseLong(places.get(places.size() - 1).get("longitude"));
+        Double.parseDouble(places.get(places.size() - 1).get("longitude"));
 
-    latitude2 = Long.parseLong(places.get(0).get("latitude"));
-    longitude2 = Long.parseLong(places.get(0).get("longitude"));
+    latitude2 = Double.parseDouble(places.get(0).get("latitude"));
+    longitude2 = Double.parseDouble(places.get(0).get("longitude"));
 
     int distance =
         computeDistance(latitude1, latitude2, longitude1, longitude2);
@@ -59,19 +59,19 @@ public class DistancesRequest extends Request {
     return distances;
   }
 
-  private int computeDistance(long latitude1, long latitude2,
-                              long longitude1, long longitude2) {
-    latitude1 = (long)Math.toRadians((double)latitude1);
-    latitude2 = (long)Math.toRadians((double)latitude2);
-    longitude1 = (long)Math.toRadians((double)longitude1);
-    longitude2 = (long)Math.toRadians((double)longitude2);
+  private int computeDistance(double latitude1, double latitude2,
+                              double longitude1, double longitude2) {
+    latitude1 = Math.toRadians(latitude1);
+    latitude2 = Math.toRadians(latitude2);
+    longitude1 = Math.toRadians(longitude1);
+    longitude2 = Math.toRadians(longitude2);
 
     double radius = getEarthRadius();
-    long longitudeDifference = longitude2 - longitude1;
+    double longitudeDifference = longitude2 - longitude1;
     double distance = radius * Math.acos(
       Math.sin(latitude1) * Math.sin(latitude2) +
         Math.cos(latitude1) * Math.cos(latitude2) * Math.cos(longitudeDifference));
-    return (int)Math.round((double)distance);
+    return (int)Math.round(distance);
   }
 
   public void setEarthRadius(double earthRadius) {
